Fixed x64 build: fixed broken menus; added General.CoolMenuEnable option to advanced options dialog; fixed cast warnings.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@7686 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/DlgFilePreview.cpp
@@ -378,7 +378,7 @@ BOOL CFilePreviewDlg::RunManual(HANDLE hFile)
 
 	m_nRange = m_nPosition = 0;
 
-	for ( QWORD nRange = 0 ; nRange < m_pRanges.GetSize() ; nRange += 2 )
+	for ( QWORD nRange = 0 ; nRange < (QWORD)m_pRanges.GetSize() ; nRange += 2 )
 	{
 		m_nRange += m_pRanges.GetAt( nRange + 1 );
 	}
@@ -387,7 +387,7 @@ BOOL CFilePreviewDlg::RunManual(HANDLE hFile)
 
 	BYTE* pData = new BYTE[ BUFFER_SIZE ];
 
-	for ( QWORD nRange = 0 ; nRange < m_pRanges.GetSize() ; nRange += 2 )
+	for ( QWORD nRange = 0 ; nRange < (QWORD)m_pRanges.GetSize() ; nRange += 2 )
 	{
 		QWORD nOffset = m_pRanges.GetAt( nRange );
 		QWORD nLength = m_pRanges.GetAt( nRange + 1 );
